[pulumi] refactor: Pulumiコンポーネントのディレクトリ構造を整理
- 各コンポーネントを専用ディレクトリに分離
  - env-config-loader/
  - github-repo/
  - lambda-deployment-bucket/
  - ssm-parameter-helper/

- 型定義をtypes.tsファイルに分離してコードの責務を明確化

- tsconfig.jsonにパスマッピングを追加
  - @tielec-iac/components: メインエクスポート
  - @tielec-iac/components/*: 個別コンポーネント

- 不要なエントリポイントファイルを削除してディレクトリ構造をクリーンに

Fixes #223
