+++ pulumi/components/src/env-config-loader/types.ts
@@ -1,0 +1,3 @@
+export interface EnvConfig {
+  [key: string]: any;
+}
