+++ pulumi/components/src/index.ts
@@ -1,27 +1,45 @@
+// GitHub Repo コンポーネント
 export {
   GitHubRepoCheckout,
+} from "./github-repo/index";
+export type {
   GitHubRepoCheckoutArgs,
-} from "./github-repo";
+} from "./github-repo/types";
 
+// Lambda Packager コンポーネント
 export {
   LambdaPackage,
+  PackageResult,
+  DEFAULT_RUNTIME,
+  DEFAULT_EXCLUDE_PATTERNS,
+} from "./lambda-packager/index";
+export type {
   LambdaPackageArgs,
-} from "./lambda-packager";
+} from "./lambda-packager/types";
 
+// Lambda Deployment Bucket コンポーネント
 export {
   LambdaDeploymentBucket,
+} from "./lambda-deployment-bucket/index";
+export type {
   LambdaDeploymentBucketArgs,
-} from "./lambda-deployment-bucket";
+} from "./lambda-deployment-bucket/types";
 
+// Env Config Loader コンポーネント
 export {
   EnvConfigLoader,
-  EnvConfig,
   loadEnvConfig,
-} from "./env-config-loader";
+} from "./env-config-loader/index";
+export type {
+  EnvConfig,
+} from "./env-config-loader/types";
 
+// SSM Parameter Helper コンポーネント
 export {
   SSMParameterHelper,
+  createSSMParameter,
+} from "./ssm-parameter-helper/index";
+export type {
   SSMParameterHelperArgs,
   SSMParameterType,
-  createSSMParameter,
-} from "./ssm-parameter-helper";
+} from "./ssm-parameter-helper/types";
